validaciones finalizadas

+++ src/main/java/com/agricola/backend/controllers/RegistroFitosanitarioRestController.java
@@ -1,9 +1,13 @@
 package com.agricola.backend.controllers;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -13,7 +17,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.agricola.backend.models.entity.RegistroFitosanitario;
@@ -28,30 +31,61 @@ public class RegistroFitosanitarioRestController {
 	private IRegistroFitosanitarioService registroFitosanitarioService;
 
 	@Secured({ "ROLE_ADMIN", "ROLE_ENCARGADOBPA","ROLE_ADMINCAMPO" })
-	@GetMapping("/registrosFitosanitarios") // okei
+	@GetMapping("/registrosFitosanitarios")
 	public List<RegistroFitosanitario> listaFitosanitarios() {
 		return registroFitosanitarioService.findAll();
 	}
 
 	@Secured({ "ROLE_ADMIN", "ROLE_ENCARGADOBPA" })
-	@GetMapping("/registrosFitosanitarios/{id}") // okei
-	public RegistroFitosanitario buscarRegistroFitosanitarioById(@PathVariable Long id) {
-		return registroFitosanitarioService.findById(id);
+	@GetMapping("/registrosFitosanitarios/{id}")
+	public ResponseEntity<?> buscarRegistroFitosanitarioById(@PathVariable Long id) {
+		
+		RegistroFitosanitario registroFito = null;
+		Map<String, Object> response = new HashMap<>();
+		
+		try {
+			registroFito = registroFitosanitarioService.findById(id);
+		} catch (DataAccessException e) {
+			response.put("mensaje", "Error al realizar la consulta en la base de datos");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		if (registroFito == null) {
+			response.put("mensaje", "El registro fitosanitario con ID : "+ id +" no existe en la base de datos");
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<RegistroFitosanitario>(registroFito, HttpStatus.OK);
 	}
 
 	@Secured({ "ROLE_ADMIN", "ROLE_ENCARGADOBPA" })
 	@PostMapping("/registrosFitosanitarios")
-	@ResponseStatus(HttpStatus.CREATED) // okei
-	public RegistroFitosanitario crearRegistroFitosanitario(@RequestBody RegistroFitosanitario registroFitosanitario) {
-		return registroFitosanitarioService.save(registroFitosanitario);
+	public ResponseEntity<?> crearRegistroFitosanitario(@RequestBody RegistroFitosanitario registroFitosanitario) {
+		
+		Map<String, Object> response = new HashMap<>();
+		
+		RegistroFitosanitario registroFito = null;
+		
+		try {
+			registroFito = registroFitosanitarioService.save(registroFitosanitario);
+		} catch (DataAccessException e) {
+			response.put("mensaje", "Error al guardar el registro fitosanitario");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		response.put("mensaje", "¡El registro fitosanitario ha sido creado con éxito!");
+		response.put("encargado", registroFito);
+		
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
 	}
 
 	@Secured({ "ROLE_ADMIN", "ROLE_ENCARGADOBPA" })
 	@PutMapping("/registrosFitosanitarios/{id}")
-	@ResponseStatus(HttpStatus.CREATED) // okei
-	public RegistroFitosanitario actualizarFitosanitario(@RequestBody RegistroFitosanitario registroFitosanitario,
-			@PathVariable Long id) {
+	public ResponseEntity<?> actualizarFitosanitario(@RequestBody RegistroFitosanitario registroFitosanitario, @PathVariable Long id) {
+		
+		Map<String, Object> response = new HashMap<>();
 
+		RegistroFitosanitario registroFitoActualizado = null;
 		RegistroFitosanitario registroFitosanitarioActual = registroFitosanitarioService.findById(id);
 
 		registroFitosanitarioActual.setTipoMaquinaria(registroFitosanitario.getTipoMaquinaria());
@@ -61,16 +95,42 @@ public class RegistroFitosanitarioRestController {
 		registroFitosanitarioActual.setHoraTermino(registroFitosanitario.getHoraTermino());
 		registroFitosanitarioActual.setCondicionesMetereologicas(registroFitosanitario.getCondicionesMetereologicas());
 
-		return registroFitosanitarioService.save(registroFitosanitarioActual);
+		try {
+			registroFitoActualizado = registroFitosanitarioService.save(registroFitosanitarioActual);
+		} catch (DataAccessException e) {
+			response.put("mensaje", "Error al actualizar el registro fitosanitario en la base de datos");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
+		response.put("mensaje", "¡El registro fitosanitario ha sido actualizado con éxito!");
+		response.put("registroFitosanitario", registroFitoActualizado);
+
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
+
 
 	}
 
 	@Secured({ "ROLE_ADMIN", "ROLE_ENCARGADOBPA" })
 	@DeleteMapping("/registrosFitosanitarios/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT) // okei
-	public void eliminarRegistroFitosanitario(@PathVariable Long id) {
+	public ResponseEntity<?> eliminarRegistroFitosanitario(@PathVariable Long id) {
 
-		registroFitosanitarioService.delete(id);
+		Map<String, Object> response = new HashMap<>();
+
+		if (registroFitosanitarioService.findById(id) == null) {
+			response.put("mensaje", "Error, el registro fitosanitario a eliminar no existe en la base de datos");
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+		}
+
+		try {
+			registroFitosanitarioService.delete(id);
+		} catch (DataAccessException e) {
+			response.put("mensaje", "Error al eliminar el registro fitosanitario en la base de datos");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		response.put("mensaje", "El registro fitosanitario fue eliminado con éxito");
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.OK);
 	}
 
 }
